fix: Fix SEGFAULT in BLS Model Loading
This change contains the minimal change to avoid SEGFAULT failures during the BLS Model Loading test.

The crash itself is cause by deleting a shared-memory region's control allocation which can happen
when we somehow endup with handle{1} (the control region) in our accounting, and then delete it when
its refcount reaches zero.

This change does fix the root cause of how we're accounting for handle{1} (which we should never have).